+++ DSP2803x_examples_ccsv5/i2c_eeprom/User/User_Globals.h
@@ -228,7 +228,7 @@ extern volatile float Voltage_minimum;
 extern volatile float SOH;
 extern volatile Uint16 Cycles;
 
-extern volatile int Pre_Charge_Measure;
+extern volatile int16 Pre_Charge_Measure;
 extern volatile int Proximity_Measure;
 extern volatile int Pilot_Measure;
